# openapi.yaml
openapi: 3.1.0
info:
  title: OpenFamHub API
  version: "1.0.0"
  description: |
    Read-mostly aggregation API for the OpenFamHub family command center. Aggregates
    external sources (Todoist, iCal, Mealie, a Monarch Money sidecar, Open-Meteo)
    behind one small backend so credentials never reach the browser, plus a
    small first-party economy (chore points -> reward redemption).

    Almost every read endpoint returns **503** when its source isn't configured
    (missing token/URL/etc.) rather than an error — the frontend keeps rendering
    from bundled demo data in that case. See `GET /api/health` for a
    per-source configured/not-configured summary.

    Data-source endpoints (tasks, calendar, meals, budget, weather, photos) are
    served by whichever provider is selected in `server/providers/registry.js`;
    paths/shapes below are the *contract* the active provider must satisfy, not
    tied to any one vendor.
  license:
    name: MIT
servers:
  - url: /api
...
paths:
...
  /rewards/{id}/assign:
    post:
      tags: [rewards]
      summary: Assign a catalog reward to a member's "working toward" list
      parameters: [{ $ref: "#/components/parameters/RewardId" }]
      requestBody: { $ref: "#/components/requestBodies/MemberId" }
      responses:
        "200": { $ref: "#/components/responses/Ok" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
  /rewards/{id}/unassign:
    post:
      tags: [rewards]
      summary: Remove a reward from a member's "working toward" list
      parameters: [{ $ref: "#/components/parameters/RewardId" }]
      requestBody: { $ref: "#/components/requestBodies/MemberId" }
      responses:
        "200": { $ref: "#/components/responses/Ok" }
        "400": { $ref: "#/components/responses/BadRequest" }
  /rewards/{id}/redeem:
    post:
      tags: [rewards]
      summary: Redeem a reward — deducts balance, logs, drops the assignment
      parameters: [{ $ref: "#/components/parameters/RewardId" }]
      requestBody: { $ref: "#/components/requestBodies/MemberId" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties: { ok: { type: boolean }, balance: { type: integer } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: Insufficient balance
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
  /journal:
    get:
      tags: [journal]
      summary: Family journal feed/timeline + the "on this day" callback + tag catalog
      description: >-
        Public/no-auth — the wall is login-free by design, same as every
        other tap-to-complete surface. Both the wall (Feed/Timeline views)
        and the PWA read this same endpoint.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  entries:
                    type: array
                    items: { $ref: "#/components/schemas/JournalEntry" }
                  onThisDay:
                    description: Past-year entries sharing today's month-day
                    type: array
                    items: { $ref: "#/components/schemas/JournalEntry" }
                  tags:
                    type: array
                    items: { $ref: "#/components/schemas/JournalTag" }
    post:
      tags: [journal]
      summary: Create a journal entry
      description: >-
        PWA-only — nothing is authored on-wall by design. The author is
        always the logged-in session, never the request body.
      security: [{ sessionCookie: [] }]
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              required: [text]
              properties:
                text: { type: string, maxLength: 2000 }
                tag: { type: string, description: "One of the tag catalog's ids, or omitted" }
                memberIds: { type: string, description: "JSON-encoded array of member ids this entry is about" }
                photo: { type: string, format: binary }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties: { ok: { type: boolean }, entry: { $ref: "#/components/schemas/JournalEntry" } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /journal/{id}/edit:
    post:
      tags: [journal]
      summary: Edit an entry — self-service, only the original poster
      security: [{ sessionCookie: [] }]
      parameters: [{ $ref: "#/components/parameters/JournalEntryId" }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [text]
              properties:
                text: { type: string, maxLength: 2000 }
                tag: { type: string, nullable: true }
                memberIds: { type: array, items: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties: { ok: { type: boolean }, entry: { $ref: "#/components/schemas/JournalEntry" } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404":
          description: Entry doesn't exist, or isn't this session's own
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /journal/{id}/delete:
    post:
      tags: [journal]
      summary: Delete an entry — self-service, only the original poster. Cleans up its photo file, if any.
      security: [{ sessionCookie: [] }]
      parameters: [{ $ref: "#/components/parameters/JournalEntryId" }]
      responses:
        "200": { $ref: "#/components/responses/Ok" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404":
          description: Entry doesn't exist, or isn't this session's own
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /journal/{id}/heart:
    post:
      tags: [journal]
      summary: Tap-to-heart — anonymous, like every other on-wall tap (no login there by design)
      parameters: [{ $ref: "#/components/parameters/JournalEntryId" }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties: { ok: { type: boolean }, hearts: { type: integer } }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /mobile/login:
    post:
      tags: [mobile]
      summary: Companion PWA login (per-member 4-digit PIN)
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [memberId, pin]
              properties:
                memberId: { type: string }
                pin: { type: string }
      responses:
        "200":
          description: OK — sets an HttpOnly session cookie
          content:
            application/json:
              schema:
                type: object
                properties: { memberId: { type: string } }
        "401":
          description: Wrong PIN, unknown member, or locked out
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
  /mobile/me:
    get:
      tags: [mobile]
      summary: The logged-in member for the current session cookie
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties: { memberId: { type: string } }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /mobile/chores:
    get:
      tags: [mobile]
      summary: The logged-in member's own Routine·Chore tasks + balance
      security: [{ sessionCookie: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  tasks:
                    type: array
                    items: { $ref: "#/components/schemas/Task" }
                  balance: { type: integer }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "503": { $ref: "#/components/responses/NotConfigured" }
  /mobile/chores/{id}/{action}:
    post:
      tags: [mobile]
      summary: Close/reopen one of the logged-in member's own chores
      security: [{ sessionCookie: [] }]
      parameters:
        - $ref: "#/components/parameters/TaskId"
        - $ref: "#/components/parameters/CloseReopenAction"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties: { ok: { type: boolean }, balance: { type: integer } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "502": { $ref: "#/components/responses/UpstreamError" }
  /push/vapid-key:
    get:
      tags: [push]
      summary: The server's VAPID public key (not a secret)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  key: { type: string }
        "503": { $ref: "#/components/responses/NotConfigured" }
  /push/status:
    get:
      tags: [push]
      summary: Whether push is configured and this member has a subscription
      security: [{ sessionCookie: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  configured: { type: boolean }
                  subscribed: { type: boolean }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /push/subscribe:
    post:
      tags: [push]
      summary: Register this browser's push subscription for the logged-in member
      description: >-
        Body is the browser's PushSubscription.toJSON() ({ endpoint, keys:
        { p256dh, auth } }). Scheduled notifications (chore reminder, dinner
        digest) go to the members listed in PUSH_MEMBERS. Requires a secure
        context on the client — see docs/deployment/self-hosting.md for
        `tailscale serve` HTTPS mode.
      security: [{ sessionCookie: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [endpoint, keys]
              properties:
                endpoint: { type: string }
                keys:
                  type: object
                  required: [p256dh, auth]
                  properties:
                    p256dh: { type: string }
                    auth: { type: string }
      responses:
        "200": { $ref: "#/components/responses/Ok" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /push/unsubscribe:
    post:
      tags: [push]
      summary: Drop a push subscription by endpoint
      security: [{ sessionCookie: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [endpoint]
              properties:
                endpoint: { type: string }
      responses:
        "200": { $ref: "#/components/responses/Ok" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: openfamhub_session
  parameters:
...
    TaskId:
      name: id
      in: path
      required: true
      schema: { type: string }
    RewardId:
      name: id
      in: path
      required: true
      schema: { type: string }
    CloseReopenAction:
      name: action
      in: path
      required: true
      schema: { type: string, enum: [close, reopen] }
    JournalEntryId:
      name: id
      in: path
      required: true
      schema: { type: integer }
  requestBodies:
    MemberId:
      content:
        application/json:
          schema:
            type: object
            required: [memberId]
            properties: { memberId: { type: string } }
  responses:
    Ok:
      description: OK
      content:
        application/json:
          schema:
            type: object
            properties: { ok: { type: boolean } }
    BadRequest:
      description: Invalid input
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Unknown resource
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: Not logged in
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    NotConfigured:
      description: Source not configured, or upstream unreachable — client falls back to demo data
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    UpstreamError:
      description: Upstream (e.g. Todoist) rejected the write
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
...
    Error:
      type: object
      properties:
        error: { type: string }
...
    Task:
      type: object
      properties:
        id: { type: string }
        title: { type: string }
        done: { type: boolean }
...
    JournalTag:
      type: object
      properties:
        id: { type: string, enum: [milestone, quote, school, health] }
        label: { type: string }
        icon: { type: string }
        catKey: { type: string }
    JournalEntry:
      type: object
      properties:
        id: { type: integer }
        authorId: { type: string, description: "Who posted it (the PWA session)" }
        text: { type: string }
        tag: { type: string, nullable: true }
        photoPath: { type: string, nullable: true, description: "Filename served at /api/journal/img/{photoPath}" }
        hearts: { type: integer }
        localDate: { type: string, format: date, description: "Server-local date, drives timeline grouping + on-this-day matching" }
        createdAt: { type: string, format: date-time }
        memberIds:
          description: Who this entry is about/involves (separate from authorId)
          type: array
          items: { type: string }
...